refactor(frame): split address classification out of inspect.rs
inspect.rs crossed the 2000-LOC [REFACTOR] tier (2097). This moves the
seven address-classification functions to addr_class.rs, taking it to
1878.

The group was chosen by measurement rather than by line count. It is the
one cluster in the file with ZERO internal dependencies: it calls nothing
else defined in inspect.rs, so the move carries the whole property
instead of scattering it. The remaining clusters -- header/offset
walking, fragment classification, session-key parsing -- do depend on
each other, and splitting any of them would put one property across two
files, which is not a seam.

v4_addr_is_directed_broadcast has no callers outside the group and moves
with the functions it serves.

inspect.rs re-exports the six pub names rather than leaving callers to be
rewritten, because the tree reaches them by BOTH shapes: a bare use
(icmp_ptb.rs) and a fully-qualified
crate::afxdp::frame::inspect::l2_dst_is_group_or_broadcast (tcp.rs). The
second would break silently on a move, and a re-export keeps the diff to
the two files that matter.

No functional change: the bodies are moved verbatim.
